fix(chapter): fix chapter members comparison bug
Fix a bug in the `FOSS Chapter` doctype where the comparison of chapter members was not done correctly. This bug caused issues when checking if a member was already a team member or not; resulting in role removal of all the members of chapter. Now, the comparison is done using the `chapter_member` attribute of each member, ensuring accurate results.