chore(chapter): pass  user from client side in `check_if_chapter_member` method
There is a bug where the user is met with an error dialog with a message "You are not the part of this chapter". This happens due to discrepancy in session user in client side and server side. The easiest way to fix this is to send the session user from the client side. This commit essentially does that for the required method