FEAT: Use and display Matrix public chat URL on city chapter page
The changes in this commit were performed manually i.e. i did not launch
the frappe app and rename the field via that. I simply updated the text
via the editor.

On the dashboard, the matrix link can now be provided, which should
automatically show up on the city page